Make parallel-downloads scenario represent a more realistic use case
It's not clear what use case is represented by wanting the location
of another project's artifact, but not the artifact itself _and_ not
considering that information part of the inputs of the task.  The most
obvious case would be IDE metadata generation, but this is handled
differently anyways (i.e. there is explicit locking controlling
access to the other project's state).  Changing this to represent
more realistic tasks.